job_aeh/requirements.txt
azure-eventhub
azure-eventhub-checkpointstoreblob
azure-identity
python-dotenv
